examples/gemma3/requirements.txt
datasets
onnx==1.18.0
onnx-ir==0.1.4
onnxruntime-genai-cuda==0.9.0
onnxruntime-gpu==1.22.0
onnxscript==0.3.2
optimum
setuptools
tabulate
transformers
